refactor: reorganize styles directory structure
