+++ src/styles/common/NextBtn.scss
@@ -1,6 +1,6 @@
 // 다음 페이지 이동 버튼 스타일
 
-@import './root.scss';
+@import '../root.scss';
 
 .move-btn {
   cursor: pointer;
